refactor: rename moll.metrics to moll.measures
