--- moll/metrics/__init__.py
@@ -1,1 +1,0 @@
-from ._metrics import *  # noqa: F403 (import *)
